requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg==3.1.19
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
